fix(YoutubeSponsorBlock): add missing categories and split plugin files
