+++ src/equicordplugins/youtubeSponsorBlock.desktop/style.css
@@ -90,6 +90,10 @@
     background: #008fd6;
 }
 
+.vc-youtube-sponsorblock-hook {
+    background: #ff1684;
+}
+
 .vc-youtube-sponsorblock-filler {
     background: #7300ff;
 }
